Add additional support for Church programs in demo application.
Switch the set of active examples depending on which perspective is active (HOGM or Church). Add Church examples based on those in the Design and Implementation Document. Drop the 'Import...' menu item, instead a sub-menu item is to be added to the 'Window' menu item that will let the user select between the HOGM (default) and Church perspectives. Church lexer to support scheme multi-line comments, i.e. #| comments |#. Church translation logic to no longer capture throwables but passes them on so they can be output in the GUI as well.

git-svn-id: https://aic-praise.googlecode.com/svn/trunk@1281 e74e62f3-1872-5a1b-a7a8-c7220c51b9cb

Former-commit-id: eff4ed95618985c9311769b64565b5bd1925da99