Improvements to Demo UI.
Rule converter now explicitly checks for random variables in the query that do not exist in the model and throws an explicit exception if any identified. UI will now automatically switch from the problems tab (if selected) to the results tab when a successful query is run. The query combo box now no longer expands incorrectly when large queries are entered.

git-svn-id: https://aic-praise.googlecode.com/svn/trunk@1325 e74e62f3-1872-5a1b-a7a8-c7220c51b9cb

Former-commit-id: a3e2ac3a8d1d38d6faa3fd48bf9aeeaa8bb1146b