Added tests for the new implementation of a tabular BayesianNode
The tests of NewTableBayesianNode are in NewTableBayesianModel.java. It
is better coded and more memory friendly (using "entries from
TableFactor" to save the parameters instead of duplicating with another
data structure) but in fact slower than the original one. Maybe it would
be good to work a bit more to improve its speed, or continue with the
original one.